requirements.txt
# Google REST client apis
google-api-python-client
google-auth

# Cloud client apis
google-cloud-storage
google-cloud-logging

# SSH keys
cryptography
gitpython

# Configuration templating
Jinja2

# Displaying YAML/Reading expanded deployment configs
pyyaml

# Python Natural Language Toolkit-b3 dictionary attack
nltk
